+++ packages/sdk/src/order/fill-order/index.ts
@@ -18,6 +18,7 @@ import type { CheckAssetTypeFunction } from "../check-asset-type"
 import { checkAssetType } from "../check-asset-type"
 import { checkLazyAssetType } from "../check-lazy-asset-type"
 import { checkChainId } from "../check-chain-id"
+import type { IRaribleEthereumSdkConfig } from "../../types"
 import type {
 	CryptoPunksOrderFillRequest,
 	FillOrderAction,
@@ -32,6 +33,8 @@ import { OpenSeaOrderHandler } from "./open-sea"
 import { CryptoPunksOrderHandler } from "./crypto-punks"
 import type { OrderFillTransactionData, FillOrderStageId } from "./types"
 import type { OrderFillSendData } from "./types"
+import type { GetOrderBuyTxRequest } from "./types"
+import type { TransactionData } from "./types"
 
 export class OrderFiller {
 	v1Handler: RaribleV1OrderHandler
@@ -47,12 +50,14 @@ export class OrderFiller {
 		private readonly config: EthereumConfig,
 		private readonly apis: RaribleEthereumApis,
 		private readonly getBaseOrderFee: (type: SimpleOrder["type"]) => Promise<number>,
+		private readonly sdkConfig?: IRaribleEthereumSdkConfig
 	) {
 		this.getBaseOrderFillFee = this.getBaseOrderFillFee.bind(this)
 		this.getTransactionData = this.getTransactionData.bind(this)
+		this.getBuyTx = this.getBuyTx.bind(this)
 		this.v1Handler = new RaribleV1OrderHandler(ethereum, apis.order, send, config, getBaseOrderFee)
 		this.v2Handler = new RaribleV2OrderHandler(ethereum, send, config, getBaseOrderFee)
-		this.openSeaHandler = new OpenSeaOrderHandler(ethereum, send, config, apis, getBaseOrderFee)
+		this.openSeaHandler = new OpenSeaOrderHandler(ethereum, send, config, apis, getBaseOrderFee, sdkConfig)
 		this.punkHandler = new CryptoPunksOrderHandler(ethereum, send, config, getBaseOrderFee)
 		this.checkAssetType = checkAssetType.bind(this, apis.nftCollection)
 		this.checkLazyAssetType = checkLazyAssetType.bind(this, apis.nftItem)
@@ -102,6 +107,22 @@ export class OrderFiller {
 	 * Accept bid order
 	 */
 	acceptBid: FillOrderAction = this.getFillAction()
+
+	async getBuyTx({request, from}: GetOrderBuyTxRequest): Promise<TransactionData> {
+		const inverted = await this.invertOrder(request, from)
+		if (request.assetType && inverted.make.assetType.assetClass === "COLLECTION") {
+			inverted.make.assetType = await this.checkAssetType(request.assetType)
+		}
+		const {functionCall, options} = await this.getTransactionRequestData(request, inverted)
+		const callInfo = await functionCall.getCallInfo()
+		const value = options.value?.toString() || "0"
+		return {
+			from,
+			value,
+			data: functionCall.data,
+			to: callInfo.contract,
+		}
+	}
 
 	private async invertOrder(request: FillOrderRequest, from: Address) {
 		switch (request.order.type) {
@@ -195,6 +216,7 @@ export class OrderFiller {
 			inverted.make.assetType = await this.checkAssetType(request.assetType)
 		}
 		const {functionCall, options} = await this.getTransactionRequestData(request, inverted)
+
 		return {
 			data: functionCall.data,
 			options,