+++ packages/sdk/src/types.ts
@@ -1,4 +1,5 @@
 import type { ConfigurationParameters } from "@rarible/ethereum-api-client"
+import type { Word } from "@rarible/types"
 
 export type EthereumNetwork = "e2e" | "ropsten" | "rinkeby" | "mainnet" | "mumbai" | "mumbai-dev" | "polygon" | "dev-ethereum" | "dev-polygon"
 
@@ -11,4 +12,10 @@ export enum LogsLevel {
 export interface IRaribleEthereumSdkConfig {
 	apiClientParams?: ConfigurationParameters
 	logs?: LogsLevel
+	ethereum?: EthereumNetworkConfig
+	polygon?: EthereumNetworkConfig
+}
+
+export interface EthereumNetworkConfig {
+	openseaOrdersMetadata?: Word
 }